@@ -2,37 +2,50 @@ source 'https://rubygems.org'
 
 
 gem 'pg'
-gem 'rails', '3.2.11'
+# gem 'rails', '3.2.11'
 
-group :assets do
-  gem 'sass-rails',   '~> 3.2.6'
-  gem 'coffee-rails', '~> 3.2.1'
-  gem 'uglifier', '>= 1.0.3'
-end
+gem "rails", '~>4.0.0'
+gem 'dalli'
+gem 'simple_form', git: 'git://github.com/plataformatec/simple_form.git'
+gem 'sass-rails'
+gem 'coffee-rails'
+gem 'uglifier'
+gem 'haml'
+gem 'devise', '3.1.1'
+gem 'cancan'
+gem 'cancan_strong_parameters' # allows cancan to use rails4
+gem 'protected_attributes' # allows rails4 to use attr_accessible
+
+gem 'protected_attributes'
+gem 'activeresource', github: 'rails/activeresource'
+gem 'actionpack-action_caching', github: 'rails/actionpack-action_caching'
+gem 'actionpack-page_caching', github: 'rails/actionpack-page_caching'
+gem 'activerecord-session_store'
+gem 'rails-observers'
+gem 'actionview-encoded_mail_to'
+gem 'rails-perftest'
+gem 'actionpack-xml_parser', github: 'rails/actionpack-xml_parser'
+
 
 group :development do
   gem "haml-rails"
-  gem "hpricot", ">= 0.8.6"
-  gem "ruby_parser", ">= 3.1.1"
-  gem "quiet_assets", ">= 1.0.1"
-  gem "better_errors", ">= 0.3.2"
-  gem "binding_of_caller", ">= 0.6.8"
+  gem "hpricot"
+  gem "ruby_parser"
+  gem "quiet_assets"
+  gem "better_errors"
+  gem "binding_of_caller"
   gem "meta_request"
   gem 'awesome_print'
   gem 'pry'
   # gem 'sqlite3'
   gem 'pg'
-  gem "rspec-rails", ">= 2.12.2"
-  gem "factory_girl_rails", ">= 4.1.0"
+  gem "rspec-rails"
+  gem "factory_girl_rails"
 end
 
 gem 'jquery-rails'
 gem 'jquery-ui-rails'
-gem "haml"
-gem "devise", ">= 2.2.2"
-gem "cancan", ">= 1.6.8"
 gem "rolify", ">= 3.2.0"
-gem "simple_form"
 gem "paperclip"
 gem 'aws-sdk'
 gem 'ruby-mp3info', :require => 'mp3info'
