@@ -9,27 +9,30 @@
 # from scratch. The latter is a flawed and unsustainable approach (the more migrations
 # you'll amass, the slower it'll run and the greater likelihood for issues).
 #
-# It's strongly recommended to check this file into your version control system.
+# It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(:version => 20130716160821) do
+ActiveRecord::Schema.define(version: 20131104002524) do
 
-  create_table "addresses", :force => true do |t|
+  # These are extensions that must be enabled in order to support this database
+  enable_extension "plpgsql"
+
+  create_table "addresses", force: true do |t|
     t.string   "line1"
     t.string   "line2"
     t.string   "city"
     t.string   "state"
     t.string   "zip"
     t.integer  "user_id"
-    t.datetime "created_at", :null => false
-    t.datetime "updated_at", :null => false
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
   end
 
-  create_table "albums", :force => true do |t|
+  create_table "albums", force: true do |t|
     t.string   "title"
     t.string   "artist"
     t.integer  "year"
-    t.datetime "created_at",                                                            :null => false
-    t.datetime "updated_at",                                                            :null => false
+    t.datetime "created_at",                                                     null: false
+    t.datetime "updated_at",                                                     null: false
     t.string   "cover_art_file_name"
     t.string   "cover_art_content_type"
     t.integer  "cover_art_file_size"
@@ -39,16 +42,17 @@ ActiveRecord::Schema.define(:version => 20130716160821) do
     t.integer  "category_id"
     t.integer  "album_type"
     t.integer  "catalog_id"
-    t.integer  "quantity",                                             :default => 0,   :null => false
-    t.decimal  "price",                  :precision => 7, :scale => 2, :default => 0.0, :null => false
+    t.integer  "quantity",                                       default: 0,     null: false
+    t.decimal  "price",                  precision: 7, scale: 2, default: 0.0,   null: false
+    t.boolean  "digital",                                        default: false, null: false
   end
 
-  create_table "blog_posts", :force => true do |t|
+  create_table "blog_posts", force: true do |t|
     t.string   "title"
     t.text     "post"
     t.string   "slug"
-    t.datetime "created_at",              :null => false
-    t.datetime "updated_at",              :null => false
+    t.datetime "created_at",              null: false
+    t.datetime "updated_at",              null: false
     t.string   "blog_cover_file_name"
     t.string   "blog_cover_content_type"
     t.integer  "blog_cover_file_size"
@@ -56,78 +60,78 @@ ActiveRecord::Schema.define(:version => 20130716160821) do
     t.string   "soundcloud_url"
   end
 
-  create_table "categories", :force => true do |t|
+  create_table "categories", force: true do |t|
     t.string   "name"
     t.text     "description"
-    t.datetime "created_at",  :null => false
-    t.datetime "updated_at",  :null => false
+    t.datetime "created_at",  null: false
+    t.datetime "updated_at",  null: false
   end
 
-  create_table "order_items", :force => true do |t|
+  create_table "order_items", force: true do |t|
     t.integer  "product_id"
     t.string   "product_type"
     t.integer  "order_id"
-    t.integer  "quantity",     :default => 0
-    t.datetime "created_at",                  :null => false
-    t.datetime "updated_at",                  :null => false
+    t.integer  "quantity",     default: 0
+    t.datetime "created_at",               null: false
+    t.datetime "updated_at",               null: false
   end
 
-  create_table "orders", :force => true do |t|
+  create_table "orders", force: true do |t|
     t.integer  "user_id"
     t.string   "status"
-    t.datetime "created_at", :null => false
-    t.datetime "updated_at", :null => false
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
     t.integer  "address_id"
   end
 
-  create_table "profile_types", :force => true do |t|
+  create_table "profile_types", force: true do |t|
     t.string   "name"
     t.text     "description"
-    t.datetime "created_at",  :null => false
-    t.datetime "updated_at",  :null => false
+    t.datetime "created_at",  null: false
+    t.datetime "updated_at",  null: false
   end
 
-  create_table "roles", :force => true do |t|
+  create_table "roles", force: true do |t|
     t.string   "name"
     t.integer  "resource_id"
     t.string   "resource_type"
-    t.datetime "created_at",    :null => false
-    t.datetime "updated_at",    :null => false
+    t.datetime "created_at",    null: false
+    t.datetime "updated_at",    null: false
   end
 
-  add_index "roles", ["name", "resource_type", "resource_id"], :name => "index_roles_on_name_and_resource_type_and_resource_id"
-  add_index "roles", ["name"], :name => "index_roles_on_name"
+  add_index "roles", ["name", "resource_type", "resource_id"], name: "index_roles_on_name_and_resource_type_and_resource_id", using: :btree
+  add_index "roles", ["name"], name: "index_roles_on_name", using: :btree
 
-  create_table "songs", :force => true do |t|
+  create_table "songs", force: true do |t|
     t.string   "title"
     t.string   "artist"
     t.integer  "length"
     t.integer  "size"
     t.integer  "album_id"
-    t.datetime "created_at",                      :null => false
-    t.datetime "updated_at",                      :null => false
+    t.datetime "created_at",                   null: false
+    t.datetime "updated_at",                   null: false
     t.string   "mp3_file_name"
     t.string   "mp3_content_type"
     t.integer  "mp3_file_size"
     t.datetime "mp3_updated_at"
     t.integer  "track"
     t.text     "metadata"
-    t.integer  "downloads",        :default => 0, :null => false
+    t.integer  "downloads",        default: 0, null: false
   end
 
-  create_table "users", :force => true do |t|
-    t.string   "email",                  :default => "", :null => false
-    t.string   "encrypted_password",     :default => "", :null => false
+  create_table "users", force: true do |t|
+    t.string   "email",                  default: "", null: false
+    t.string   "encrypted_password",     default: "", null: false
     t.string   "reset_password_token"
     t.datetime "reset_password_sent_at"
     t.datetime "remember_created_at"
-    t.integer  "sign_in_count",          :default => 0
+    t.integer  "sign_in_count",          default: 0
     t.datetime "current_sign_in_at"
     t.datetime "last_sign_in_at"
     t.string   "current_sign_in_ip"
     t.string   "last_sign_in_ip"
-    t.datetime "created_at",                             :null => false
-    t.datetime "updated_at",                             :null => false
+    t.datetime "created_at",                          null: false
+    t.datetime "updated_at",                          null: false
     t.string   "name"
     t.string   "avatar_file_name"
     t.string   "avatar_content_type"
@@ -137,14 +141,14 @@ ActiveRecord::Schema.define(:version => 20130716160821) do
     t.integer  "account_request"
   end
 
-  add_index "users", ["email"], :name => "index_users_on_email", :unique => true
-  add_index "users", ["reset_password_token"], :name => "index_users_on_reset_password_token", :unique => true
+  add_index "users", ["email"], name: "index_users_on_email", unique: true, using: :btree
+  add_index "users", ["reset_password_token"], name: "index_users_on_reset_password_token", unique: true, using: :btree
 
-  create_table "users_roles", :id => false, :force => true do |t|
+  create_table "users_roles", id: false, force: true do |t|
     t.integer "user_id"
     t.integer "role_id"
   end
 
-  add_index "users_roles", ["user_id", "role_id"], :name => "index_users_roles_on_user_id_and_role_id"
+  add_index "users_roles", ["user_id", "role_id"], name: "index_users_roles_on_user_id_and_role_id", using: :btree
 
 end
